app/requirements.txt
numpy
uvicorn
fastapi
pillow
tensorflow
python-multipart
python-dotenv
